feat: match table, landscape orientation on match screen and multiplayer socket logic to context api
