+++ package.json
@@ -28,6 +28,7 @@
     "axios": "^1.13.2",
     "expo": "~54.0.9",
     "expo-constants": "~18.0.10",
+    "expo-screen-orientation": "^55.0.16",
     "expo-secure-store": "~15.0.7",
     "expo-splash-screen": "~31.0.10",
     "expo-status-bar": "~3.0.8",
