+++ src/types/multiplayer.types.ts
@@ -8,23 +8,77 @@ export type LobbyPlayer = {
   isEliminated?: boolean;
 };
 
-export type HandCard = {
-  id: string;
-  type: string;
-  color: string;
-};
+// ─── Partida e Componentes do Jogo ───────────────────────────────────────────
 
-// ─── Sala ─────────────────────────────────────────────────────────────────────
+export enum CardType {
+  Essence = 'essence',
+  Joker = 'joker',
+  BlackHole = 'black_hole',
+  Vortex = 'vortex',
+  BuyPlus1 = 'buy_plus_1',
+  BuyPlus2 = 'buy_plus_2',
+  StealNextOne = 'steal_next_1',
+  StealPrevOne = 'steal_prev_1',
+  StealAnyOne = 'steal_any_1',
+  Trap = 'trap',
+  Recycle = 'recycle',
+  BlockPurchase = 'block_purchase',
+  SwapNextHand = 'swap_next_hand',
+  SwapPrevHand = 'swap_prev_hand',
+  SwapAnyHand = 'swap_any_hand',
+  ExtraPower = 'extra_power',
+  BlockSteal = 'block_steal',
+  Reflect = 'reflect',
+  Nullify = 'nullify',
+}
+
+export enum GamePhase {
+  Idle = 'idle',
+  Draw = 'draw',
+  Play = 'play',
+  Resolve = 'resolve',
+  React = 'react',
+  End = 'end',
+}
+
+export enum TurnDirection {
+  Clockwise = 'clockwise',
+  CounterClockwise = 'counter_clockwise',
+}
+
+export interface Card {
+  id: string;
+  type: CardType;
+  color: string;
+}
+
+export interface PlayerInfo {
+  id: string;
+  socketId: string;
+  name: string;
+  avatar?: string;
+  isEliminated: boolean;
+  canReturn: boolean;
+  hand: Card[] | { count: number };
+}
+
 export type RoomPhase = 'waiting' | 'playing' | 'finished';
 
-// ─── Partida ──────────────────────────────────────────────────────────────────
-export type GameState = {
+export interface PlayerView {
   roomId: string;
-  currentTurnPlayerId: string;
-  players: LobbyPlayer[];
-  playerHands: Record<string, HandCard[]>;
-  turnNumber?: number;
-};
+  players: PlayerInfo[];
+  deck: { remaining: number };
+  discardPile: Card[];
+  currentTurnIndex: number;
+  direction: TurnDirection;
+  turnNumber: number;
+  phase: GamePhase;
+  pendingInterrupt: any | null;
+  pendingDiscard: any | null;
+  blockPurchaseTurnsRemaining: number;
+  purchaseBlockedThisTurn: boolean;
+  hasPlayedCardThisTurn: boolean;
+}
 
 // ─── Payloads do servidor ─────────────────────────────────────────────────────
